Add arbitrary attributes to NetCDF fields
This update extends diag_manager to give the user the ability to add
an arbitrary number of attributes to a diagnostic field variable.

This update initially limits the number of attributes to 2, but
this can be adjusted via a new namelist variable max_field_attributes.

The attribute can have any arbitrary name, and its value can be a
1D array of integers or reals, or a character string.