Remote rregion from history filename
The reason for appending rregion to the diagnostic file name is unknown at this point, and a reason cannot be found for keeping this around. Also, adding rregion to the file name, and then having FRE remove this string later on is fragile. An example of this, is the addition diag_manager prepending the date to the file name (instead of FRE) broke the section in the run script that removed rregion from the file name.

The addition of rregion needs to be removed from diag_manager.

Fixes #11