Changed order of opertaions in register_static_field to fix a problem with diag field log.
Somewhare between verona and pre-warsaw, a call to find_input_field
and RETURN statement were introduced in register_static_field. Since this
check was done at the very beginning of the subroutine, it rendered the
diag field logging facility practically useless: only the fields already
in the diag_table would be reported in the logi whose purpose was to list
all fields.

log_diag_field_info shold be called unconditionally for the diag log to
work properly; this change moves find_input_field check below for this
purpose.