Field manager and fm_util can cast integer values to real.
Before this change, the result of the following input in the field_table:
    array1 = 0.5e-01,   1
was unexpectedly (and silently) weird. Because the second value is interpreted
by field manager parser as integer, the whole array was reset and the only
element was set to integer 1. Current mod makes the field manager understand
that the entire array is real.

Also, it allows function fm_util_get_real to return real values even if the field
manager stored the input value is integer-typed. E.g. in the following example:
   real_factor = 4
previous code would crash when calling fm_util_get_real; this code returns
expected value of 4.0

NOTE that this typecast only happens from integers to reals; mixture of numbers
and strings (or logicals) in input arrays in field_table is still handled the
same way as before.