+Added 3 new coupler_type interfaces
  Added the new interfaces coupler_type_destructor, coupler_type_increment_data,
and coupler_type_rescale_data. Also changed the interfaces to coupler_type_spawn
to accomodate arrays with halos and copy over the fields flux_type and
implementation, which had previously been omitted.  New optional arguments were
added to coupler_type_copy_data to select a single field or work on a specified
halo size and to permit a 2d type to be copied to a 3d type.  Answers do not
change in any test cases that do not use these new interfaces, and these
routines have now all been tested in a full coupled model implementation.