+Added coupler_type_initialized
  Added a new variable to the coupler_types indicating whether a particular
coupler_bc_type variable has already been initialized, along with a new
function, coupler_type_initialized, that can be used to query the value of this
variable.  In addition, coupler_type_spawn has a new optional argument,
as_needed, that will only cause a couper type to be spawned if it has not been
initialized but its parent has.  All answers are bitwise identical, but there
are interface changes.