Update .travis.yml for new MPI directive
The previous travis environment used an outdated openMPI the
did not have a new MPI API element.  The travis enviornment
now uses xenial with an openMPI library with the missing
API.

Other travis.yml cleanup was also done.

Also found that fm_util.o file was not included in the final
libFMS library.  This has been corrected.

The base test did not link in the library, added a no-op
libFMS call to constants_init(), and modified the test
Makefile.am to link in the library.