Remove Cray Enhanced IO functionality
The Cray Enhanced IO functionality is not supported on our Cray system, and as such has not been tested.  This commit removes (as much as possible) the bits that would `call ASSIGN` to configure the enhanced IO functionality.

This leaves the call to open_ieee32_file, as it is used in other model components, but is now no more than a _special_ mpp_open.