Zhi nest merge required test fixes (#1)
* Correct syntax error in test_common.sh.in

This commit removes the line in test_common.sh.in that had a syntax error.  This removal was done to allow test_common.sh.in to support more MPI run wrappers than the openMPI wrapper.

* Correct additional syntax error in test_common.sh.in

I introduced another syntax error.  This corrects that.

* Remove html requiement from diag_integral Makefile

This is needed to allow the zhi_nest_merge to pass the Travis CI tests