Explict string length for buffer in axis_utils
The `buffer` character array in axis_utils was incorrectly sorted in the
`select type` blocks in PGI due to an inability to match the
assumed-length character type.  This behavior differs from Intel and
GNU.

This patch uses the same fix to the char array by passing a slice of the
whole array (`buffer(1:128)`).  I can only assume this signals to the
compiler that it is assumed-length.

This resolves observed issues in the OM_1deg MOM6 regression test.