mpp_type_free cleanup
This patch fixes some issues and errors in the `mpp_type_free` funcrion,
mostly related to the handling of the default `mpp_byte` type.

MPI datatypes are handled through the `mpp_type` wrapper for
non-contiguous data transfers within collectives.  User-defined types
are stored in a linked list, with a minimal wrapper to `MPI_BYTE` stored
at the head.  Because this type is intrinsic, it is handled separately.

The cleanup function `mpp_type_free` had some errors where it was not
cleaning up these types, and (had it been working) would have
incorrectly tried to de-register `MPI_BYTE`, which causes a SIGSEGV.

This patch resolves the linked list iteration and also excepts the
de-register and deallocation associated with `MPI_BYTE`.

This also has the very small benefit of fixing up some minor memory
leaks associated with `mpp_byte`.